refactor: Page 컴포넌트에서 Header과 Content 컴포넌트로 분리
