+++ src/components/ProductContent/ProductContent.styles.ts
@@ -1,0 +1,19 @@
+import { css } from '@emotion/css';
+
+export const productWrapper = css`
+  padding: 20px;
+`;
+
+export const productPageTitle = css`
+  font-size: 24px;
+  font-weight: 700;
+  line-height: 100%;
+  margin: 10px 0px;
+`;
+
+export const selectBoxContainer = css`
+  display: flex;
+  gap: 150px;
+  justify-content: space-between;
+  margin: 20px 0px;
+`;
