+++ src/pages/ProductsPage/ProductsPage.style.ts
@@ -5,21 +5,3 @@ export const productPageContainer = css`
   margin: 0 auto;
   padding-top: 64px;
 `;
-
-export const productWrapper = css`
-  padding: 20px;
-`;
-
-export const productPageTitle = css`
-  font-size: 24px;
-  font-weight: 700;
-  line-height: 100%;
-  margin: 10px 0px;
-`;
-
-export const selectBoxContainer = css`
-  display: flex;
-  gap: 150px;
-  justify-content: space-between;
-  margin: 20px 0px;
-`;
